refactor: 같은 슬롯일땐 순차처리 아닐때는 병렬처리로 큐 구조 개선
