--- src/main/java/roomescape/infra/queue/AsyncMessage.java
@@ -1,4 +1,0 @@
-package roomescape.infra.queue;
-
-public record AsyncMessage<T>(String jobId, T request) {
-}
