+++ src/test/java/roomescape/domain/waiting/WaitingQueueTest.java
@@ -1,27 +1,37 @@
 package roomescape.domain.waiting;
 
+import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.awaitility.Awaitility.await;
 import static org.junit.jupiter.api.Assertions.assertAll;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.when;
 
 import java.time.LocalDate;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
 import roomescape.domain.waiting.dto.WaitingRequest;
 import roomescape.domain.waiting.dto.WaitingResponse;
-import roomescape.infra.queue.AsyncMessage;
+import roomescape.exception.ErrorCode;
+import roomescape.exception.RoomescapeException;
 import roomescape.infra.queue.JobResult;
 
+@ExtendWith(MockitoExtension.class)
 class WaitingQueueTest {
 
+    @Mock
+    private WaitingService waitingService;
+
     private WaitingQueue waitingQueue;
-    private WaitingRequest request;
 
     @BeforeEach
     void setUp() {
-        waitingQueue = new WaitingQueue();
-        request = new WaitingRequest("유저1", LocalDate.of(2099, 12, 31), 1L, 1L);
+        waitingQueue = new WaitingQueue(waitingService);
     }
 
     @Nested
@@ -29,60 +39,84 @@ class WaitingQueueTest {
     class Enqueue {
 
         @Test
-        void enqueue_하면_PENDING_상태로_등록된다() {
+        void enqueue_하면_즉시_PENDING_상태로_등록된다() {
+            WaitingRequest request = new WaitingRequest("유저1", LocalDate.of(2099, 12, 31), 1L, 1L);
+
             String jobId = waitingQueue.enqueue(request);
 
             JobResult result = waitingQueue.getResult(jobId);
-
             assertAll(
                     () -> assertThat(jobId).isNotNull(),
-                    () -> assertThat(result.status()).isEqualTo("PENDING"),
-                    () -> assertThat(result.data()).isNull(),
-                    () -> assertThat(result.errorMessage()).isNull()
-            );
-        }
-
-        @Test
-        void enqueue_하면_큐에서_take로_꺼낼_수_있다() throws InterruptedException {
-            waitingQueue.enqueue(request);
-
-            AsyncMessage message = waitingQueue.take();
-
-            assertAll(
-                    () -> assertThat(message.request()).isEqualTo(request),
-                    () -> assertThat(message.jobId()).isNotNull()
+                    () -> assertThat(result.status()).isEqualTo("PENDING")
             );
         }
     }
 
     @Nested
-    @DisplayName("storeResult 테스트")
-    class StoreResult {
+    @DisplayName("처리 성공 케이스")
+    class ProcessSuccess {
 
         @Test
-        void storeResult_하면_getResult로_저장된_결과를_반환한다() {
-            String jobId = waitingQueue.enqueue(request);
+        void createWaiting_성공_시_SUCCESS_상태와_응답이_저장된다() {
+            WaitingRequest request = new WaitingRequest("유저1", LocalDate.of(2099, 12, 31), 1L, 1L);
             WaitingResponse response = new WaitingResponse(1L, "유저1", LocalDate.of(2099, 12, 31), 1L, 1L);
+            when(waitingService.createWaiting(any())).thenReturn(response);
 
-            waitingQueue.storeResult(jobId, JobResult.success(response));
+            String jobId = waitingQueue.enqueue(request);
+
+            await().atMost(1, SECONDS).until(() -> !"PENDING".equals(waitingQueue.getResult(jobId).status()));
 
             JobResult result = waitingQueue.getResult(jobId);
             assertAll(
                     () -> assertThat(result.status()).isEqualTo("SUCCESS"),
-                    () -> assertThat(result.data()).isEqualTo(response)
+                    () -> assertThat(result.data()).isEqualTo(response),
+                    () -> assertThat(result.errorMessage()).isNull()
             );
         }
+    }
+
+    @Nested
+    @DisplayName("처리 실패 케이스")
+    class ProcessFailure {
 
         @Test
-        void storeResult_실패_결과를_저장하면_FAILED_상태와_에러_메시지를_반환한다() {
+        void RoomescapeException_발생_시_FAILED_상태와_에러_메시지가_저장된다() {
+            WaitingRequest request = new WaitingRequest("유저1", LocalDate.of(2099, 12, 31), 1L, 1L);
+            when(waitingService.createWaiting(any()))
+                    .thenThrow(new RoomescapeException(ErrorCode.DUPLICATE_WAITING_NAME));
+
             String jobId = waitingQueue.enqueue(request);
 
-            waitingQueue.storeResult(jobId, JobResult.failed("에러 발생"));
+            await().atMost(1, SECONDS).until(() -> !"PENDING".equals(waitingQueue.getResult(jobId).status()));
 
             JobResult result = waitingQueue.getResult(jobId);
             assertAll(
                     () -> assertThat(result.status()).isEqualTo("FAILED"),
-                    () -> assertThat(result.errorMessage()).isEqualTo("에러 발생")
+                    () -> assertThat(result.errorMessage()).isEqualTo(ErrorCode.DUPLICATE_WAITING_NAME.getMessage()),
+                    () -> assertThat(result.data()).isNull()
+            );
+        }
+
+        @Test
+        void 같은_슬롯의_요청은_순서대로_처리된다() {
+            WaitingRequest request1 = new WaitingRequest("유저1", LocalDate.of(2099, 12, 31), 1L, 1L);
+            WaitingRequest request2 = new WaitingRequest("유저2", LocalDate.of(2099, 12, 31), 1L, 1L);
+            WaitingResponse response1 = new WaitingResponse(1L, "유저1", LocalDate.of(2099, 12, 31), 1L, 1L);
+            WaitingResponse response2 = new WaitingResponse(2L, "유저2", LocalDate.of(2099, 12, 31), 1L, 1L);
+            when(waitingService.createWaiting(request1)).thenReturn(response1);
+            when(waitingService.createWaiting(request2)).thenReturn(response2);
+
+            String jobId1 = waitingQueue.enqueue(request1);
+            String jobId2 = waitingQueue.enqueue(request2);
+
+            await().atMost(1, SECONDS).until(() ->
+                    !"PENDING".equals(waitingQueue.getResult(jobId1).status()) &&
+                    !"PENDING".equals(waitingQueue.getResult(jobId2).status())
+            );
+
+            assertAll(
+                    () -> assertThat(waitingQueue.getResult(jobId1).status()).isEqualTo("SUCCESS"),
+                    () -> assertThat(waitingQueue.getResult(jobId2).status()).isEqualTo("SUCCESS")
             );
         }
     }
@@ -93,9 +127,7 @@ class WaitingQueueTest {
 
         @Test
         void 등록되지_않은_jobId면_null을_반환한다() {
-            JobResult result = waitingQueue.getResult("없는-job-id");
-
-            assertThat(result).isNull();
+            assertThat(waitingQueue.getResult("없는-job-id")).isNull();
         }
     }
 
@@ -107,6 +139,7 @@ class WaitingQueueTest {
         void 기준일_이전_날짜의_결과는_제거된다() {
             WaitingRequest pastRequest = new WaitingRequest("유저1", LocalDate.of(2000, 1, 1), 1L, 1L);
             String pastJobId = waitingQueue.enqueue(pastRequest);
+            await().atMost(1, SECONDS).until(() -> !"PENDING".equals(waitingQueue.getResult(pastJobId).status()));
 
             waitingQueue.evictExpiredResults();
 
@@ -117,6 +150,7 @@ class WaitingQueueTest {
         void 오늘_날짜의_결과는_제거되지_않는다() {
             WaitingRequest todayRequest = new WaitingRequest("유저1", LocalDate.now(), 1L, 1L);
             String todayJobId = waitingQueue.enqueue(todayRequest);
+            await().atMost(1, SECONDS).until(() -> !"PENDING".equals(waitingQueue.getResult(todayJobId).status()));
 
             waitingQueue.evictExpiredResults();
 
@@ -125,7 +159,9 @@ class WaitingQueueTest {
 
         @Test
         void 미래_날짜의_결과는_제거되지_않는다() {
-            String futureJobId = waitingQueue.enqueue(request);
+            WaitingRequest futureRequest = new WaitingRequest("유저1", LocalDate.of(2099, 12, 31), 1L, 1L);
+            String futureJobId = waitingQueue.enqueue(futureRequest);
+            await().atMost(1, SECONDS).until(() -> !"PENDING".equals(waitingQueue.getResult(futureJobId).status()));
 
             waitingQueue.evictExpiredResults();
 
@@ -135,8 +171,13 @@ class WaitingQueueTest {
         @Test
         void 과거와_미래_결과가_섞여있으면_과거만_제거된다() {
             WaitingRequest pastRequest = new WaitingRequest("유저1", LocalDate.of(2000, 1, 1), 1L, 1L);
+            WaitingRequest futureRequest = new WaitingRequest("유저2", LocalDate.of(2099, 12, 31), 1L, 1L);
             String pastJobId = waitingQueue.enqueue(pastRequest);
-            String futureJobId = waitingQueue.enqueue(request);
+            String futureJobId = waitingQueue.enqueue(futureRequest);
+            await().atMost(1, SECONDS).until(() ->
+                    !"PENDING".equals(waitingQueue.getResult(pastJobId).status()) &&
+                    !"PENDING".equals(waitingQueue.getResult(futureJobId).status())
+            );
 
             waitingQueue.evictExpiredResults();
